cruds 80%, aun no se puede editar ninguno

+++ app/Http/Controllers/Admin/EmployeeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Employee;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class EmployeeController extends Controller
 {
@@ -24,7 +25,12 @@ class EmployeeController extends Controller
     }
     public function update(Request $request, $id)
     {
-        $datosEmpleado=$request->except(["_token", '_method']); 
+        $datosEmpleado=$request->except(["_token", '_method']);
+        if($request->hasFile('photo')){
+            $employee=Employee::findOrFail($id);
+            Storage::delete('public/'.$employee->photo);
+            $datosEmpleado['photo']=$request->file('photo')->store('uploads', 'public');
+        }
         Employee::where('id', '=',$id)->update($datosEmpleado);
 
         $employee=Employee::findOrFail($id);
@@ -38,13 +44,19 @@ class EmployeeController extends Controller
             }
            
             Employee::insert($datosEmpleado);
-            return response()->json($datosEmpleado);
+            //return response()->json($datosEmpleado);
+            return redirect('admin/empleados')->with('mensaje', 'Empleado creado exitosamente');
+        
     }
     public function destroy($id)
     {
         //
-        Employee::destroy($id);
-     return redirect('admin/empleados');
+        $employee=Employee::findOrFail($id);
+        if( Storage::delete('public/'.$employee->photo)){
+            Employee::destroy($id);
+        }
+        
+     return redirect('admin/empleados')->with('mensaje', 'Empleado eliminado exitosamente');;
 
     }
