+++ app/Http/Controllers/Admin/ServiceController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Service;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ServiceController extends Controller
 {
@@ -15,20 +16,25 @@ class ServiceController extends Controller
     }
     public function create()
     {
-        return view ('admin.servicios.create');
+        $service = new Service();
+        return view ('admin.servicios.create', compact('service'));
     }
     public function edit($id)
     {
-        $datosServicio=Service::findOrFail($id);
+        $service=Service::findOrFail($id);
         return view ('admin.servicios.edit', compact('service'));
     }
     public function update(Request $request, $id)
     {
         $datosServicio=$request->except(["_token", '_method']); 
+        if($request->hasFile('photo')){
+            $service=Service::findOrFail($id);
+            Storage::delete('public/'.$service->photo);
+            $datosServicio['photo']=$request->file('photo')->store('uploads', 'public');
+        }
         Service::where('id', '=',$id)->update($datosServicio);
-
-        $datosServicio=Service::findOrFail($id);
-        return view ('admin.empleados.edit', compact('employee'));
+        $service=Service::findOrFail($id);
+        return view ('admin.servicios.edit', compact('service'));
     }
     public function store(Request $request)
     {
@@ -38,13 +44,18 @@ class ServiceController extends Controller
             }
            
             Service::insert($datosServicio);
-            return response()->json($datosServicio);
+            //return response()->json($datosServicio);
+            return redirect('admin/servicios')->with('mensaje', 'Servicio creado exitosamente');
     }
     public function destroy($id)
     {
         //
-        Service::destroy($id);
-     return redirect('admin/servicios');
+        $service=Service::findOrFail($id);
+        if( Storage::delete('public/'.$service->photo)){
+            Service::destroy($id);
+        }
+        
+     return redirect('admin/servicios')->with('mensaje', 'Servicio eliminado exitosamente');;
 
     }
 }
